Add useOutsideClick hook
